player attack && grass effect health 1 HP

+++ dungeons_plugin/src/resources/tile.rs
@@ -10,7 +10,8 @@ pub enum Tile {
     Spawn, // 出生点
     Grass, // 草地单位
     Enemy(EnemyType), // 敌方单位
-    EnemyNeighbor(u8), // 敌方临近单位
+    /// 临近格展示值：邻格敌方单位 **HP 之和**（生成期来自类型推算，运行期与实体 [`crate::components::entity_status::Health`] 同步）。
+    EnemyNeighbor(u16),
     Treasure, // 宝藏
     OutWay, // 出口
     Safe, // 安全点
